use github-authentication instead
