--- extensions/authentication-proxy/.vscodeignore
@@ -1,7 +1,0 @@
-src/**
-test/**
-out/**
-tsconfig.json
-build/**
-extension.webpack.config.js
-yarn.lock
