--- extensions/authentication-proxy/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "authentication-proxy",
-  "displayName": "%displayName%",
-  "description": "%description%",
-  "publisher": "vscode",
-  "license": "MIT",
-  "version": "1.0.0",
-  "engines": {
-    "vscode": "^1.5.0"
-  },
-  "aiKey": "AIF-d9b70cd4-b9f9-4d70-929b-a071c400b217",
-  "categories": [
-    "Other"
-  ],
-  "activationEvents": [
-    "onAuthenticationRequest:github"
-  ],
-  "extensionDependencies": [
-    "vscode.git-base"
-  ],
-  "main": "./out/extension",
-  "icon": "resources/icons/git.png",
-  "scripts": {
-    "compile": "gulp compile-extension:git",
-    "watch": "gulp watch-extension:git",
-    "test": "node ../../node_modules/mocha/bin/mocha"
-  },
-  "capabilities": {
-    "virtualWorkspaces": true,
-    "untrustedWorkspaces": {
-      "supported": true
-    }
-  },
-  "contributes": {
-    "commands": [
-      {
-        "command": "authentication.login.device.code",
-        "title": "Get Device Code",
-        "category": "Authentication Proxy"
-      }
-    ]
-  },
-  "dependencies": {
-    "node-fetch": "2.6.1"
-  },
-  "devDependencies": {
-    "@types/mocha": "^8.2.0",
-    "@types/node": "14.x",
-    "@types/node-fetch": "^2.5.7"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/vscode.git"
-  }
-}
