--- extensions/authentication-proxy/package.nls.json
@@ -1,4 +1,0 @@
-{
-	"displayName": "Authentication Proxy",
-	"description": "Authentication Proxy"
-}
