--- extensions/authentication-proxy/src/test/extension.test.ts
@@ -1,13 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import 'mocha';
-import * as assert from 'assert';
-
-suite('authentication-proxy', () => {
-	test('empty test', () => {
-		assert.deepStrictEqual([], []);
-	});
-});
