--- extensions/authentication-proxy/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"extends": "../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "./out",
-		"experimentalDecorators": true,
-		"typeRoots": [
-			"./node_modules/@types"
-		]
-	},
-	"include": [
-		"src/**/*",
-		"../../src/vscode-dts/vscode.d.ts",
-		"../types/lib.textEncoder.d.ts"
-	]
-}
